Create output range like interface in Logger
This way any logger can choose to implement logHeader, logMsgPart
and finishLogMsg which can be implemented without memory allocation
for some loggers. If the logger does not overwrite these functions
they construct a string with an appender and call writeLogMsg
just like before this commit.

This is just as efficient as std.string.format, as std.string.format
internally uses appender + formattedWrite as well.

MsgRange is used as a thin adapter which any reasonable compiler
should remove/inline. The reason is that we might have to introduce
another OutputRange interface for structured logging so
we can't have two put methods with potentially the same parameter
types.