Port filelogger, stdiologger to Output Range API
Avoid memory allocation.

Right now only implemented if the default logging format is used.
In theory the API can work with any format string. We need to figure
out what part of the header needs to be logged before the message
and log it in logHeader, then log the message, then log the rest of the
header in finishLogMsg.